Optimize tensor batch operations
Keep pre-batch data buffers in RAM until the batch is required, in which
case the data batch will be prepared in-RAM and sent all at once to the
GPU. This should reduce the amount of tensors kept on the GPU in order
to make the affected modules (batch predict and replay buffer) more
scalable.

+++ src/game/experience/Experience.ts
@@ -16,7 +16,7 @@ export interface Experience {
     /** Resultant state from action. */
     nextState: Float32Array[];
     /** Binary choice legality data for the {@link nextState next state}. */
-    choices: Int32Array;
+    choices: Uint8Array;
     /** Marks {@link nextState} as a terminal state so it won't be processed. */
     done: boolean;
 }
